# backend/docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.ProjectListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Project'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  models.ProjectReview:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      effectiveDate:
        type: string
      enforcementStatusId:
        type: integer
      enforcementStatusName:
        type: string
      id:
        type: integer
      nameEnglish:
        type: string
      nameThai:
        type: string
      ownerGroupId:
        type: integer
      ownerGroupName:
        type: string
      remarks:
        type: string
      review_cancel_summary_remark:
        description: วันที่ยกเลิกสรุปผลการเวียน
        type: string
      review_circulation_summary_remark:
        description: วันที่สรุปผลการเวียน
        type: string
      reviewEndDate:
        type: string
      reviewStartDate:
        type: string
      stageCode:
        type: string
      stageUiMsg:
        type: string
      tisNumber:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: integer
    type: object
  models.ProjectReviewListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.ProjectReview'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  models.ProjectReviewLog:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      id:
        type: integer
      projectReviewId:
        type: integer
      remarks:
        type: string
      stageCode:
        type: string
      stageDate:
        type: string
      stageDescription:
        type: string
      stageStatus:
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/estandard/project-review:
    get:
      consumes:
      - application/json
      description: Get list of project reviews with search and pagination
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Stage code filter
        in: query
        name: stageCode
        type: string
      - description: Enforcement status filter
        in: query
        name: enforcementStatus
        type: string
      - description: Owner group ID filter
        in: query
        name: ownerGroupId
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectReviewListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Projects Review List
      tags:
      - Projects
  /api/v1/estandard/project-review/{id}:
    put:
      consumes:
      - application/json
      description: Update a project review by ID
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project review data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProjectReview'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectReview'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Project Review
      tags:
      - Projects
  /api/v1/estandard/project-review/{id}/logs:
    get:
      consumes:
      - application/json
      description: Get all log entries for a project review
      parameters:
      - description: Project Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProjectReviewLog'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Project Review Logs
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create or update a project review log entry
      parameters:
      - description: Project Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project review log data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProjectReviewLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectReviewLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upsert Project Review Log
      tags:
      - Projects
  /api/v1/estandard/projects:
    get:
      consumes:
      - application/json
      description: Get list of projects with search and pagination
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Start year filter
        in: query
        name: startYear
        type: string
      - description: Stage code filter
        in: query
        name: stageCode
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Projects List
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create or update a project
      parameters:
      - description: Project data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upsert Project
      tags:
      - Projects
  /api/v1/estandard/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Get a single project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Project by ID
      tags:
      - Projects
  /api/v1/estandard/projects/{id}/logs:
    get:
      consumes:
      - application/json
      description: Get all logs for a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProjectLog'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Project Logs
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create or update a project log entry
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project log data
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/models.ProjectLog'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upsert Project Log
      tags:
      - Projects
  /api/v1/estandard/projects/{id}/save-standard-announcement:
    post:
      consumes:
      - application/json
      description: Save standard announcement data to both i_projects and tb3_tis
        tables
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Standard announcement data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SaveStandardAnnouncementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Save Standard Announcement
      tags:
      - Projects
  /api/v1/estandard/projects/check-tis-number:
    get:
      consumes:
      - application/json
      description: Check if a TIS number already exists in tb3_tis table
      parameters:
      - description: TIS Number to check
        in: query
        name: tisNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response with exists field
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check TIS Number
      tags:
      - Projects
...
swagger: "2.0"